pyproject.toml
[project]
name = "git-sync"
version = "1.0.0"
description = "Sync GitHub repos to internal Git server with author rewriting"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Git-Sync Team" }]
keywords = ["git", "sync", "author-rewrite", "github", "mirror", "tui"]

dependencies = [
    "git-filter-repo>=2.47.0",
    "apscheduler>=3.10.0",
    "click>=8.1.0",
    "textual>=0.47.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pyinstaller>=6.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
git-sync = "git_sync.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true